Working on cross-compiling for specific AVR chip
The linker can't fit the tests in the .text section.
Am I cross-compiling for the correct chip? I can't be; I'm not specifying.
A bunch of CppUTest operators are not defined. Mocks may not work...
which is fine. Why mock when we're on the real thing? Will need to rewrite stuff.