Fix bug where LED's weren't toggling properly.
Don't yet understand the solution, but the impetus came from Man3 FAQ:,
Why are (many) interrupt flags cleared by writing a logical 1?

If getting the bit number from a function, the compiler may not be able to optimize
(especially if the function is in another module?).
The |= operator results in an extra assembly operation (or, out) when compared to the = operator (out).
For some reason, |= results in inaccurate behavior, but only if optimization doesn't convert the
operation into a simple bit shift with a constant.